fix features rerendering issue
